workqueue-affinity-demo: change tunable to nice
Change the tunable workqueue parameter that is varied during the demo to
niceness since setting cpumask does not actually trigger formation of a new pool
for unbound workqueues.   Change the delays from "sleep" to "Press any key to
continue" in order to allow viewers to differentiate between the execution time
of drgn/BPF and intentional waiting.   Change the victim workqueue to
nvme-delete-wq since both laptop and PC run it.