workqueue-affinity-demo: show moving to hi-prio workqueue
When $NEWNICE is a random value 0 < $NEWNICE < 20, the kernel creates a new
kworker pool whose attributes match those specified for the workqueue.  If
NEWNICE==0 or NEWNICE=20, the existing workqueues satisfy the need.  Changing
between 0 and 20 moves the workqueue back and forth between default-niceness and
high-prio worker pools.  For other values, if there's an idle unbound worker
pool, the niceness of that pool is modified and the workqueue is assigned there.